[refactor] Update Package Name from Error to Exception
