+++ src/main/java/gp/cnusambe/service/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package gp.cnusambe.service;
 
 import gp.cnusambe.domain.User;
-import gp.cnusambe.error.UserNotFoundException;
+import gp.cnusambe.exception.UserNotFoundException;
 import gp.cnusambe.repository.UserRepository;
 import gp.cnusambe.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
