+++ src/test/java/gp/cnusambe/controller/UserControllerTest.java
@@ -1,0 +1,12 @@
+package gp.cnusambe.controller;
+
+
+import gp.cnusambe.domain.User;
+
+public class UserControllerTest {
+
+    void join(){
+        User user = new User();
+
+    }
+}
